+++ datafusion/physical-expr/src/aggregate/mod.rs
@@ -38,7 +38,6 @@ pub(crate) mod correlation;
 pub(crate) mod count;
 pub(crate) mod count_distinct;
 pub(crate) mod covariance;
-pub(crate) mod first_last;
 pub(crate) mod grouping;
 pub(crate) mod median;
 pub(crate) mod nth_value;
