+++ datafusion/physical-expr/src/expressions/mod.rs
@@ -53,7 +53,6 @@ pub use crate::aggregate::correlation::Correlation;
 pub use crate::aggregate::count::Count;
 pub use crate::aggregate::count_distinct::DistinctCount;
 pub use crate::aggregate::covariance::{Covariance, CovariancePop};
-pub use crate::aggregate::first_last::{FirstValue, LastValue};
 pub use crate::aggregate::grouping::Grouping;
 pub use crate::aggregate::median::Median;
 pub use crate::aggregate::min_max::{Max, Min};
@@ -76,6 +75,9 @@ pub use crate::window::rank::{dense_rank, percent_rank, rank};
 pub use crate::window::rank::{Rank, RankType};
 pub use crate::window::row_number::RowNumber;
 pub use crate::PhysicalSortExpr;
+pub use datafusion_aggregate_functions::first_last::{
+    FirstValuePhysicalExpr as FirstValue, LastValuePhysicalExpr as LastValue,
+};
 
 pub use binary::{binary, BinaryExpr};
 pub use case::{case, CaseExpr};
